Back completo con todos los CRUD

+++ src/main/java/com/appTest/store/dto/purchase/PurchaseDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter @Setter
@@ -13,11 +14,11 @@ public class PurchaseDTO implements Serializable {
     private Long idPurchase;
     private LocalDate datePurchase;
     private String supplierName;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
 
     public PurchaseDTO() {}
 
-    public PurchaseDTO(Long idPurchase, LocalDate datePurchase, String supplierName, Double totalAmount) {
+    public PurchaseDTO(Long idPurchase, LocalDate datePurchase, String supplierName, BigDecimal totalAmount) {
         this.idPurchase = idPurchase;
         this.datePurchase = datePurchase;
         this.supplierName = supplierName;
